Refactor dsl tests
